refactor: Create custom hook for get pairing uri
